Removed CDW support from MedicationController
