+++ data-query/src/main/java/gov/va/api/health/dataquery/service/controller/medication/Dstu2MedicationTransformer.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 import lombok.Builder;
 
 @Builder
-public class DatamartMedicationTransformer {
+public class Dstu2MedicationTransformer {
 
   private final DatamartMedication datamart;
