--- data-query/src/test/resources/cdw/old-medication-1.01.json
@@ -1,20 +1,0 @@
-{
-  "resourceType" : "Medication",
-  "id" : "2f773f73-ad7f-56ca-891e-8e364c913fe0",
-  "code" : {
-    "coding" : [
-      {
-        "system" : "https://www.nlm.nih.gov/research/umls/rxnorm/",
-        "code" : "259255",
-        "display" : "ATORVASTATIN CA 80MG TAB"
-      }
-    ],
-    "text" : "ATORVASTATIN CA 80MG TAB"
-  },
-  "product" : {
-    "id" : "4014891",
-    "form" : {
-      "text" : "TAB"
-    }
-  }
-}
