--- mr-anderson/src/test/resources/samples/Medication-1.01.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<root>
-  <fhirVersion>DSTU2 Argonaut</fhirVersion>
-  <resourceName>medication</resourceName>
-  <resourceVersion>1.01</resourceVersion>
-  <returnType>full</returnType>
-  <returnFormat>XML</returnFormat>
-  <recordsPerPage>15</recordsPerPage>
-  <pageNumber>1</pageNumber>
-  <startRecord>1</startRecord>
-  <endRecord>1</endRecord>
-  <recordCount>1</recordCount>
-  <pageCount>1</pageCount>
-  <errorNumber>0</errorNumber>
-  <errorLine>0</errorLine>
-  <medications>
-    <medication>
-      <rowNumber>1</rowNumber>
-      <cdwId>419327</cdwId>
-      <text>ASPIRIN 81MG EC TAB</text>
-      <code>
-        <text>ASPIRIN 81MG TAB,EC</text>
-        <coding>
-          <system>https://www.nlm.nih.gov/research/umls/rxnorm/</system>
-          <code>308416</code>
-          <display>ASPIRIN 81MG TAB,EC</display>
-        </coding>
-      </code>
-      <product>
-        <id>4005766</id>
-        <form>
-          <text>TAB,EC</text>
-        </form>
-      </product>
-    </medication>
-  </medications>
-</root>
